Data files and example results
Adding the data files needed to run the script (*.count.txt) , as well as an output file (ct_all.txt) and sample output plots (*.png).  The file 'ct_all.txt' can be used to generate the plots, instead of going through all the initial steps of merging the separate files.